requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
boto3==1.38.23
botocore==1.37.3
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
ffmpeg==1.4
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jmespath==1.0.1
moviepy==2.2.1
multidict==6.4.4
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.2.1
proglog==0.1.12
propcache==0.3.1
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
requests==2.32.3
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
Telethon==1.40.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
yarl==1.20.0
